ledger 핵심 기능 마무리

+++ src/main/java/com/wallet/ledger/ledger/domain/ItemDTO.java
@@ -8,17 +8,17 @@ import lombok.NoArgsConstructor;
 public class ItemDTO {
 
     protected Long id;
-    protected Long amount;
+    protected int amount;
     protected String orderId;
     protected ReferenceType type;
 
-    public ItemDTO(Long id, Long amount, String orderId) {
+    public ItemDTO(Long id, int amount, String orderId) {
         this.id = id;
         this.amount = amount;
         this.orderId = orderId;
     }
 
-    public ItemDTO(Long id, Long amount, String orderId, ReferenceType type) {
+    public ItemDTO(Long id, int amount, String orderId, ReferenceType type) {
         this.id = id;
         this.amount = amount;
         this.orderId = orderId;
