+++ src/main/java/com/wallet/ledger/ledger/domain/LedgerEntryDTO.java
@@ -9,11 +9,11 @@ import lombok.NoArgsConstructor;
 public class LedgerEntryDTO {
 
     private AccountDTO account;
-    private Long amount;
+    private Integer amount;
     private LedgerEntryType type;
 
     @Builder
-    public LedgerEntryDTO(AccountDTO account, Long amount, LedgerEntryType type) {
+    public LedgerEntryDTO(AccountDTO account, int amount, LedgerEntryType type) {
         this.account = account;
         this.amount = amount;
         this.type = type;
